refactor: Collapse remaining .skia.cs partials and renames
Finishes the suffix drop for the three Skia-only assemblies:

- FrameworkElement, AcrylicBrush and UIElement merge now that the body is
  taken up to the closing brace character. Their classes end with an
  #endregion/#endif that sits in the brace's leading trivia, so stopping
  short of it orphaned the matching opener.
- ViewExtensions.skia.cs declares Uno.UI.ViewExtensions, not the
  Uno.UI.Extensions.ViewExtensions of its same-named neighbour. Its real
  counterpart is ViewExtensions.crossruntime.cs, so it merges there.
- DamageRegionExtensions, Visual.Damage and RetainedLayer have no base
  file at all and are plain renames.

Image and DispatcherAnimator keep their suffix: their base files sit
entirely inside #if branches that are false for the only project that
compiles them, so merging would move live code into a branch that never
builds. Collapsing those needs the dead branch removed first.
