+++ src/Uno.UI.Composition/Composition/DamageRegionExtensions.cs
