+++ src/Uno.UI.Composition/Composition/Visual.Damage.cs
