+++ src/Uno.UI/UI/Xaml/FrameworkElement.cs
@@ -1,4 +1,4 @@
-﻿#pragma warning disable CS0105 // Ignore duplicate namespaces, to remove when moving to WinUI source tree.
+#pragma warning disable CS0105 // Ignore duplicate namespaces, to remove when moving to WinUI source tree.
 
 using Uno.Diagnostics.Eventing;
 using Microsoft.UI.Xaml.Controls;
@@ -31,6 +31,7 @@ using Uno.UI.Xaml.Media;
 
 using Color = System.Drawing.Color;
 using View = Microsoft.UI.Xaml.UIElement;
+using Uno.UI.Extensions;
 
 namespace Microsoft.UI.Xaml
 {
@@ -916,5 +917,96 @@ namespace Microsoft.UI.Xaml
 		{
 			(this as IDependencyObjectStoreProvider)?.Store.SetTemplatedParent2(tp);
 		}
-	}
+
+		protected FrameworkElement()
+		{
+			Initialize();
+		}
+
+		bool IFrameworkElementInternal.HasLayouter => true;
+
+		partial void Initialize();
+
+		private bool IsTopLevelXamlView() => false;
+
+		internal void SuspendRendering() => throw new NotSupportedException();
+
+		internal void ResumeRendering() => throw new NotSupportedException();
+
+		#region Name Dependency Property
+
+		private void OnNameChanged(string oldValue, string newValue)
+		{
+			if (FrameworkElementHelper.IsUiAutomationMappingEnabled)
+			{
+				Microsoft.UI.Xaml.Automation.AutomationProperties.SetAutomationId(this, newValue);
+			}
+		}
+
+		[GeneratedDependencyProperty(DefaultValue = "", ChangedCallback = true)]
+		public static DependencyProperty NameProperty { get; } = CreateNameProperty();
+
+		public string Name
+		{
+			get => GetNameValue();
+			set => SetNameValue(value);
+		}
+
+		#endregion
+
+#if DEBUG
+		private void OnGenericPropertyUpdated(DependencyPropertyChangedEventArgs args)
+		{
+		}
+#endif
+
+		private event TypedEventHandler<FrameworkElement, object> _loading;
+		public event TypedEventHandler<FrameworkElement, object> Loading
+		{
+			add
+			{
+				_loading += value;
+			}
+			remove
+			{
+				_loading -= value;
+			}
+		}
+
+		private event RoutedEventHandler _loaded;
+		public event RoutedEventHandler Loaded
+		{
+			add
+			{
+				_loaded += value;
+			}
+			remove
+			{
+				_loaded -= value;
+			}
+		}
+
+		private event RoutedEventHandler _unloaded;
+		public event RoutedEventHandler Unloaded
+		{
+			add
+			{
+				_unloaded += value;
+			}
+			remove
+			{
+				_unloaded -= value;
+			}
+		}
+
+#if ENABLE_CONTAINER_VISUAL_TRACKING // Make sure to update the Comment to have the valid depth
+		partial void OnLoading()
+		{
+			if (_visual is not null)
+			{
+				_visual.Comment = $"{this.GetDebugDepth():D2}-{this.GetDebugName()}";
+			}
+		}
+#endif
+		}
 }
